完善上传文件逻辑，重构，修改文件名

+++ src/com/ustb/ssjgl/main/dao/bean/TPotentialsFile.java
@@ -23,7 +23,7 @@ public class TPotentialsFile {
     private String cSuffix;
 
     /** 大小(单位:k) */
-    private Integer nSize;
+    private Long nSize;
 
     /** 是否有效 */
     private Integer nValid;
@@ -82,11 +82,11 @@ public class TPotentialsFile {
         this.cSuffix = cSuffix == null ? null : cSuffix.trim();
     }
 
-    public Integer getnSize() {
+    public Long getnSize() {
         return nSize;
     }
 
-    public void setnSize(Integer nSize) {
+    public void setnSize(Long nSize) {
         this.nSize = nSize;
     }
